Clippy lints and formatting
This might be nitpicky, but very often, the compilers suggestions when
it comes to lints can be beneficial for performance. E.g. using the
assign operator += instead of manually writing it gives room for
compiler optimisations. Formatting is also a smart thing to run
continually, to avoid diffs that are difficult to read.