Avoid allocation, only iterate
Iterating when simulating, instead of allocating to a vector, is usually
more performant. In this case, it is also more flexible, as the saving
procedure seems to involve some sort of iteration either way. If a
vector is needed, it is easy to write a wrapping function that collects
it.